fix import pddl4j

+++ projet_patia/src/fr/uga/pddl4j/tutorial/asp/ASP.java
@@ -9,12 +9,16 @@ import fr.uga.pddl4j.planners.ProblemFactory;
 import fr.uga.pddl4j.planners.Statistics;
 import fr.uga.pddl4j.planners.statespace.AbstractStateSpacePlanner;
 import fr.uga.pddl4j.planners.statespace.StateSpacePlanner;
+import fr.uga.pddl4j.planners.statespace.search.strategy.AStar;
+import fr.uga.pddl4j.planners.statespace.search.strategy.Node;
 import fr.uga.pddl4j.util.BitOp;
 import fr.uga.pddl4j.util.BitState;
 import fr.uga.pddl4j.util.CondBitExp;
 import fr.uga.pddl4j.util.MemoryAgent;
 import fr.uga.pddl4j.util.Plan;
 import fr.uga.pddl4j.util.SequentialPlan;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import java.io.File;
 import java.io.IOException;
@@ -114,7 +118,7 @@ public final class ASP extends AbstractStateSpacePlanner {
                             }
                         }
                         // We set the new child node information
-                        final int g = current.getCost() + 1;
+                        final double g = current.getCost() + 1;
                         if (!close.contains(next)) {
                             next.setCost(g);
                             next.setParent(current);
